#80: dal changed.

+++ dal/src/main/java/com/craftsmanship/tfm/dal/model/OrderItem.java
@@ -38,7 +38,7 @@ public class OrderItem {
         this.key = key;
     }
 
-    public long getQuantity() {
+    public int getQuantity() {
         return quantity;
     }
